Require 'plist' only on Darwin
Here we modify the location of the require for 'plist' so that we only
load the plist when the method that makes uses of it is called.  This
should remove the need to install the 'plist' gem on all systems that
make use of this module, regardless of the platform.