Ensure that we also require the users are created
Without this change, user may get created after the membership is
applied to the group.  This means that to arrive at the desired group
membership, it takes two puppet runs.

This work corrects the behavior by requiring that any users being
managed by puppet get created before the membership for the given group
is set.